Move validation code to its own package
